pyproject.toml
[tool.poetry]
name = "postingimage"
version = "0.1.0"
description = ""
authors = ["spawlov <107248409+spawlov@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
mypy = "^1.14.0"
types-requests = "^2.32.0.20241016"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
